♻️ Simplify SequenceParser class and PrioritizedChoiceParser class: inherits abstract class ReduceParser
